# swagger.yaml
openapi: 3.0.3
info:
  title: Redis Streams POC API
  description: |
    A comprehensive Node.js application demonstrating Redis Streams for real-time data processing with Express, PostgreSQL, and Sequelize.
    
    ## Features
    - 🚀 Redis Streams Integration with producer/consumer patterns
    - 📊 Consumer Groups for scalable and fault-tolerant message processing
    - 🗄️ PostgreSQL + Sequelize for persistent data storage
    - 🌐 RESTful API endpoints for triggering events
    - 📝 Complete audit trail of all events
    - 🔄 Real-time processing with event-driven architecture
    
    ## Redis Streams
    This API demonstrates various Redis Streams operations:
    - **XADD** - Publishing events to streams
    - **XREAD** - Simple event consumption
    - **XREADGROUP** - Consumer group processing
    - **XACK** - Message acknowledgment
    - **XGROUP CREATE** - Consumer group management
  version: 1.0.0
  contact:
    name: Redis Streams POC
    url: https://github.com/your-username/redis-streams-poc
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://your-domain.com
    description: Production server
...
paths:
...
  /api/users:
    get:
      tags:
        - Users
      summary: Get all users
      description: Retrieve a list of all users in the system
      parameters:
        - name: limit
          in: query
          description: Maximum number of users to return
          required: false
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          description: Number of users to skip
          required: false
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of users retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  count:
                    type: integer
                    example: 25
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - Users
      summary: Create a new user
      description: Create a new user and publish a user.created event to Redis Streams
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - email
              properties:
                username:
                  type: string
                  example: john_doe
                  description: Unique username for the user
                email:
                  type: string
                  format: email
                  example: john@example.com
                  description: User's email address
                status:
                  type: string
                  enum: [active, inactive, suspended]
                  default: active
                  description: User account status
      responses:
        '201':
          description: User created successfully and event published
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  user:
                    $ref: '#/components/schemas/User'
                  messageId:
                    type: string
                    example: "1642248000000-0"
                    description: Redis Stream message ID
                  message:
                    type: string
                    example: User created and event published successfully
        '400':
          description: Bad request - missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID
      description: Retrieve a specific user by their ID
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: string
            format: uuid
            example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  user:
                    $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Users
      summary: Update user
      description: Update a user and publish a user.updated event to Redis Streams
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: john_doe_updated
                email:
                  type: string
                  format: email
                  example: john.updated@example.com
                status:
                  type: string
                  enum: [active, inactive, suspended]
      responses:
        '200':
          description: User updated successfully and event published
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  user:
                    $ref: '#/components/schemas/User'
                  changes:
                    type: object
                    description: Object showing what fields were changed
                  messageId:
                    type: string
                    example: "1642248000000-1"
                  message:
                    type: string
                    example: User updated and event published successfully
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Users
      summary: Delete user
      description: Delete a user and publish a user.deleted event to Redis Streams
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User deleted successfully and event published
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  messageId:
                    type: string
                    example: "1642248000000-2"
                  message:
                    type: string
                    example: User deleted and event published successfully
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/users/{id}/events:
    get:
      tags:
        - Users
      summary: Get user events
      description: Get all events associated with a specific user
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          description: Maximum number of events to return
          required: false
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          description: Number of events to skip
          required: false
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: User events retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      username:
                        type: string
                        example: john_doe
                      email:
                        type: string
                        format: email
                        example: john@example.com
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/Event'
                  total:
                    type: integer
                    example: 10
                  limit:
                    type: integer
                    example: 50
                  offset:
                    type: integer
                    example: 0
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        username:
          type: string
          example: john_doe
        email:
          type: string
          format: email
          example: john@example.com
        status:
          type: string
          enum: [active, inactive, suspended]
          example: active
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
...
    Event:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        streamId:
          type: string
          example: "1642248000000-0"
          description: Redis Stream message ID
        eventType:
          type: string
          example: user.created
        payload:
          type: object
          example: { userId: "user-123", username: "john_doe" }
        processed:
          type: boolean
          example: true
        processedAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
        consumerId:
          type: string
          example: consumer-1
        error:
          type: string
          example: null
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00.000Z"
    Error:
      type: object
      properties:
        error:
          type: string
          example: User not found
        details:
          type: string
          example: The requested user ID does not exist
        path:
          type: string
          example: /api/users/123e4567-e89b-12d3-a456-426614174000
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
...